requirements.txt
pymongo
python-dotenv
prettytable
colorama
inquirer
inputimeout

